refactor: 단일 이미지 태그 등록 플로우 개선 (#44)
* refactor: 태그 최대 개수 검증 메서드 리팩토링

* refactor: TagMaxCountValidator -> TagValidator

태그 검증 관련 로직을 모두 TagValidator로 모으고, 단위 테스트를 진행했습니다.

* feat: ImageService 불필요한 의존성 제거

+++ capturecat-core/src/test/java/com/capturecat/core/DummyObject.java
@@ -10,8 +10,12 @@ import com.capturecat.core.domain.image.Image;
 
 public class DummyObject {
 
-	public static Image newMockImage(int id) {
-		return Image.builder().id((long) id).fileName("test1.jpg").fileUrl("testUrl1").build();
+	public static Image newMockImage(Long id) {
+		return Image.builder()
+			.id(id)
+			.fileName("test1.jpg")
+			.fileUrl("testUrl1")
+			.build();
 	}
 
 	public static List<Image> newMockImages(int fromId, int toId) {
